refactor: centralise harness retirement policy
The retirement map and its two consumers (rejectRetiredHarnesses in
init.go, pruneRetiredHarnesses in upgrade.go) form a single policy
spread across two files. Extracted to a dedicated
internal/cmd/harness_retirement.go so adding or removing entries is a
one-place edit. Same package, same behaviour — pure code motion.

Also expanded the pruneRetiredHarnesses doc comment to call out that
the in-memory mutation is persisted by the caller (mom upgrade's Phase
1 config.Save) — that was correct but invisible.
